+++ data-exercise/flink/src/test/java/org/mqjd/flink/jobs/chapter3/section3/MaxParallelismTest.java
@@ -34,7 +34,7 @@ public class MaxParallelismTest extends ContainerBaseTest {
             }
         }, (client, jobStatus) -> {
             if (jobStatus.equals(JobStatus.RUNNING)) {
-                TimerUtil.timeout(client::cancel, (long) 10 * 1000);
+                TimerUtil.timeout(client::cancel, 10L * 1000);
             } else if (jobStatus.isTerminalState()) {
                 result.complete(true);
             }
